Moved Attachment handler to its own file
